+++ controllers/routers/pages.js
@@ -6,7 +6,6 @@ module.exports = pages = {};
 //  DEPENDENCIES
 // ==============================
 const { _U_ } = require('../utilities/');
-const { postsModel } = require('../../models/');
 const { fetch } = require('../middleware/fetch');